backend/Django/requirements.txt
absl-py==0.11.0
asgiref==3.2.10
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.43.0
django-cors-headers==3.5.0
django-crontab==0.7.1
django-extensions==3.0.9
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
gunicorn==20.0.4
h5py==2.10.0
httplib2==0.18.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
Markdown==3.3.3
MarkupSafe==1.1.1
mysqlclient==2.0.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.44
opt-einsum==3.3.0
packaging==20.4
Pillow==8.0.1
protobuf==3.13.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scipy==1.5.3
six==1.15.0
sqlparse==0.4.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
uritemplate==3.0.1
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
